Bug fix in DialogFragment.findInstance
Line:
ft.putFragment(fragment.getArguments(), tag, fragment);

putFragment is intended to put the fragment state into a bundle for restoration at a later time. This cannot be called before a fragment has been added to the fragment manager or an exception is thrown. (case: makeIfNeed=true) From what I can tell, this is completely unnecessary and probably not intend.

Additional, added another findInstance that takes a fragment manager rather than an activity to allow it to be used with child fragment managers.