Fix Android issue 58602
If one puts fragments into action bar tabs and calls detach() on the
FragmentTransaction in the onTabUnselected callback, that transaction won't
be committed. The reason for leaving out the commit was that the
onTabSelected callback would run afterwards and do the commit but
onTabSelected runs on another TabWrapper instance so getActiveTransaction()
will create a new transaction leaving the first one uncommitted.

This is high impact since it makes action bar tabs unusable, the previous
tab content stays behind when switching to another tab.

For more details see:
https://code.google.com/p/android/issues/detail?id=58602